+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.0.1] - 2025-08-04
+## [1.0.2] - 2025-08-04
+
+### Added
+- Notification Preferences
+- Notifications
+  - Upcoming booking notification
+  - Booking end reminder
+
+## [1.0.1] - 2025-08-16
 
 ### Added
 - My bookings cancellation